requirements.txt
Django==1.4.1
South==0.7.6
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.17
celery==3.0.11
dj-database-url==0.2.1
django-celery==3.0.10
django-extensions==0.9
django-tastypie==0.9.11
gunicorn==0.14.6
ipython==0.13
kombu==2.4.7
lxml==2.3.6
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
pytz==2012f
readline==6.2.2
wsgiref==0.1.2
